correctly handle the logic of lyrics, but the timing info of strip still not quite sure

+++ structure_of_ws2812.h
@@ -15,7 +15,7 @@ typedef struct {
 			uint32_t time;
 			uint8_t red, green, blue;
 			uint8_t SPX_type;
-			uint16_t SPX_duration; // !! use ms
+			uint32_t SPX_duration;
 		} strip;
 	};
 } ws2812;
